Tag capabilities in PuppetDB to indicate the application they belong to
The components of an application and the capabilities they produced get tagged
with 'producer:APP:NAME' so that we can distinguish them by application and
application instance name beyond the usual reference by type and title. If a
component does not belong to an application instance but is mapped to a node
directly, it gets tagged as 'producer:main'.

This implements scoping of these resources at the PuppetDB level, and makes
sure that consumers only see capabilities from components of the same
application instance.

It does not fully implement scoping of components/capabilities to application
instances as the check for duplicate resources within a catalog does not take
this tag into account and only goes by type and title. IOW, if components
with the same name from different application instances get mapped to the same
node, the usual per-node duplication check will fail and generate an error.

Ultimately, scoping to an application instance should be done through a
dedicated metadata field rather than the hacky way of using tags. Tags are also
problematic since PuppetDB does not notice that a resource has changed if
only the tags for that resource change. This will likely make an innocuous
operation like renaming an application instance fail.